# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MOIL Reserve Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
export type DataModeRequestMode = typeof DataModeRequestMode[keyof typeof DataModeRequestMode];


export const DataModeRequestMode = {
  synthetic: 'synthetic',
  live: 'live',
} as const;

export interface DataModeRequest {
  mode: DataModeRequestMode;
}

export type DataSourceStatusStatus = typeof DataSourceStatusStatus[keyof typeof DataSourceStatusStatus];


export const DataSourceStatusStatus = {
  connected: 'connected',
  missing: 'missing',
  adapter_pending: 'adapter_pending',
} as const;

export interface DataSourceStatus {
  id: string;
  label: string;
  status: DataSourceStatusStatus;
  detail: string;
}

export type DataModeResponseMode = typeof DataModeResponseMode[keyof typeof DataModeResponseMode];


export const DataModeResponseMode = {
  synthetic: 'synthetic',
  live: 'live',
} as const;

export interface DataModeResponse {
  mode: DataModeResponseMode;
  live_ready: boolean;
  message: string;
  sources: DataSourceStatus[];
}

export interface HealthStatus {
  status: string;
}

export interface Mine {
  mine_id: string;
  name: string;
  district: string;
  mine_type: string;
  coordinate_status: string;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  depth_m?: number | null;
  reserve_confidence?: number;
  shortfall_probability?: number;
  dominant_driver?: string;
}

export interface ReservePoint {
  latitude: number;
  longitude: number;
  reserve_probability: number;
  spectral_score: number;
  structural_score: number;
  zone_type: string;
}

export interface HeatmapResponse {
  points: ReservePoint[];
  model_version: string;
  computed_at: string;
}

export interface ValidationMine {
  mine_id: string;
  percentile: number;
}

export interface ValidationResponse {
  confirmed_mines_checked: number;
  avg_percentile_rank: number;
  per_mine: ValidationMine[];
  method: string;
}

export interface ProductionHistory {
  date: string;
  planned_tonnage: number;
  actual_tonnage: number;
  downtime_hours: number;
  rainfall_mm: number;
  blast_delay_flag: boolean;
  data_provenance: string;
}

export interface ProductionForecast {
  mine_id: string;
  horizon_days: number;
  predicted_tonnage: number;
  planned_tonnage: number;
  shortfall_probability: number;
  dominant_driver: string;
  local_reserve_confidence: number;
  model_version: string;
  provenance: string;
}

export interface Recommendation {
  mine_id: string;
  action: string;
  detail: string;
  explanation_text: string;
  driver: string;
  generated_at: string;
}

export type GetReserveHeatmapParams = {
/**
 * Optional min longitude,min latitude,max longitude,max latitude
 */
bbox?: string;
};

export type GetProductionHistoryParams = {
/**
 * @minimum 30
 * @maximum 365
 */
days?: number;
};

export type GetProductionForecastParams = {
horizon?: GetProductionForecastHorizon;
};

export type GetProductionForecastHorizon = typeof GetProductionForecastHorizon[keyof typeof GetProductionForecastHorizon];


export const GetProductionForecastHorizon = {
  NUMBER_30: 30,
  NUMBER_60: 60,
  NUMBER_90: 90,
} as const;

export type GetMineRecommendationParams = {
horizon?: GetMineRecommendationHorizon;
};

export type GetMineRecommendationHorizon = typeof GetMineRecommendationHorizon[keyof typeof GetMineRecommendationHorizon];


export const GetMineRecommendationHorizon = {
  NUMBER_30: 30,
  NUMBER_60: 60,
  NUMBER_90: 90,
} as const;