# lib/api-zod/src/generated/types/mine.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * MOIL Reserve Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Mine {
  mine_id: string;
  name: string;
  district: string;
  mine_type: string;
  coordinate_status: string;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  depth_m?: number | null;
  reserve_confidence?: number;
  shortfall_probability?: number;
  dominant_driver?: string;
}